Add application modes
Current behavior is as follows: the program can now run in 3 different modes: client, server and hybrid. In client mode application connects to the server, in hybrid mode application creates own server and connects to itself. Server mode gives up GUI and just creates a server for other clients to connect to.